+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Update map style and legend [#710](https://github.com/azavea/echo-locator/pull/710)
 - Add frontend component library ADR [#614](https://github.com/azavea/echo-locator/pull/614)
 - Add `./scripts/ecs-shell <env>` script [#626](https://github.com/azavea/echo-locator/pull/626)
 - Bootstrap new frontend and remove old frontend [#666](https://github.com/azavea/echo-locator/pull/666)